Revert "EditKeyboardFragment refactor"
