+++ app/src/main/java/com/willowtree/vocable/settings/EditCategoriesViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.MutableLiveData
 import com.willowtree.vocable.BaseViewModel
 import com.willowtree.vocable.presets.PresetsRepository
 import com.willowtree.vocable.room.Category
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.koin.core.inject
 import java.util.*
@@ -60,10 +59,6 @@ class EditCategoriesViewModel(numbersCategoryId: String, mySayingsCategoryId: St
     fun updateCategory(category: Category) {
         backgroundScope.launch {
             presetsRepository.updateCategory(category)
-
-            liveShowCategoryAdded.postValue(true)
-            delay(CATEGORY_UPDATED_DELAY)
-            liveShowCategoryAdded.postValue(false)
         }
     }
